Restructured save file dialog, Add Escape Shortcut for Save/Get Filename Dialog
